pyproject.toml
[tool.black]
line-length = 88
target-version = ['py36']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.cache
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | __pycache__
    | _build
    | buck-out
    | build
    | dist
    | docs
    | migrations
  )/

)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = false
line_length = 88
known_first_party = ["getpaid_przelewy24"]
known_third_party = ["getpaid"]

[tool.poetry]
name = "Getpaid-Przelewy24"
version = "0.1.0"
description = "django-getpaid processor for Przelewy24"
authors = ["Konrad Lyda <konrad.lyda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django-getpaid = "^2.2.0"
requests = "^2.23.0"
swapper = "^1.1.2"
typing-extensions = "^3.7.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.1.1"
factory-boy = "^2.12.0"
black = "^19.10b0"
tox = "^3.14.5"
ipdb = "^0.13.2"
pytest = {version = "^5.3.5"}
codecov = {version = "^2.0.16"}
coverage = {version = "^5.0.3"}
requests-mock = {version = "^1.7.0"}
pytest-cov = {version = "^2.8.1"}
pytest-django = {version = "^3.8.0"}

[tool.poetry.extras]
test = ["pytest", "codecov", "coverage", "request-mock", "pytest-cov", "pytest-django"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
